refactor(classroom-attendance): Separate users columns into dedicated component
- Extract user columns logic from attendance-table.tsx
- Create new file attendance-table-users-columns.tsx for user column definitions
- Improve code organization and modularity of attendance table components
- Simplify imports and reduce complexity in main attendance table component
